Migrate block to typescript
